backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
sqlparse==0.5.3
gunicorn==21.2.0
groq==0.18.0 